+++ platform/nxp/lpc13xx/flash.c
@@ -98,11 +98,11 @@ static enum result flashGet(void *object, enum ifOption option, void *data)
   switch (option)
   {
     case IF_POSITION:
-      *(size_t *)data = interface->position;
+      *(uint32_t *)data = interface->position;
       return E_OK;
 
     case IF_SIZE:
-      *(size_t *)data = interface->size;
+      *(uint32_t *)data = interface->size;
       return E_OK;
 
     default:
@@ -120,7 +120,7 @@ static enum result flashSet(void *object, enum ifOption option,
   {
     case IF_FLASH_ERASE_SECTOR:
     {
-      const size_t address = *(const size_t *)data;
+      const uint32_t address = *(const uint32_t *)data;
 
       if (!isSectorAddressValid(interface, address))
         return E_VALUE;
@@ -139,7 +139,7 @@ static enum result flashSet(void *object, enum ifOption option,
   {
     case IF_POSITION:
     {
-      const size_t position = *(const size_t *)data;
+      const uint32_t position = *(const uint32_t *)data;
 
       if (!isPageAddressValid(interface, position))
         return E_VALUE;
